+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## [Unreleased][unreleased]
 
+- Add namespaces to Schema.from factory
+
 ## [1.3.0][] - 2021-06-25
 
 - Check schemas with references to schemas from attached models